pyproject.toml
[tool.poetry]
name = "Horaro"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.0.5"
djangorestframework = "3.13.1"
djoser = "2.1.0"
django-filter = "22.1"
gunicorn = "20.1.0"
pillow = "9.1.0"
python-dotenv = "0.20.0"
celery = "^5.2.7"
redis = "^4.5.5"
flower = "^1.2.0"
whitenoise = "^6.4.0"
grequests = "^0.6.0"
drf-yasg = "1.21.3"
django-cors-headers = "3.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-django = "^4.5.2"
flake8 = "^6.0.0"
isort = "^5.12.0"
faker = "^18.9.0"
factory-boy = "^3.2.1"
black = "^23.3.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79
py36 = true
skip-string-normalization = true
exclude = '''
  # Default exclude
  \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist

  # Custom excludes
  | manage\.py
  | \.pgdata
  | migrations
'''
